Stage -3  , Exception handling being added to out application right .

+++ src/main/java/com/example/demo/service/StudentService.java
@@ -1,5 +1,6 @@
 package com.example.demo.service;
 
+import com.example.demo.exception.StudentNotFoundException;
 import com.example.demo.model.Student;
 import com.example.demo.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +18,8 @@ public class StudentService {
     }
 
     public Student getStudentById(Long id){
-        return studentRepository.findById(id).orElse(null);
+        return studentRepository.findById(id)
+                .orElseThrow(()->new StudentNotFoundException(id));
     }
 
     public Student addStudent(Student student){
@@ -25,7 +27,8 @@ public class StudentService {
     }
 
     public Student updateStudent(Long id , Student updated) {
-        Student existing = studentRepository.findById(id).orElse(null);
+        Student existing = studentRepository.findById(id)
+                .orElseThrow(()->new StudentNotFoundException(id));
 
         if (existing != null) {
             existing.setName(updated.getName());
@@ -39,6 +42,9 @@ public class StudentService {
 
 
     public void deleteStudent(Long id){
+        if(!studentRepository.existsById(id)){
+            throw new StudentNotFoundException(id);
+        }
        studentRepository.deleteById(id);
     }
